+++ Changelog.md
@@ -1,6 +1,19 @@
 ### 🌌 Next Version
 The future is uncertain
 ___
+### 4.1.6
+Added Eltruism Datatext 3 to Retail where it will show Armor and Dodge change  
+Updated Aura borders to split Unitframe Auras and Secure Auras  
+Updated party gradient to work when not in a group  
+Updated Debuff borders to use Color by Dispel  
+Updated Debuff borders to use Elvui Colors  
+Updated shadows for some new frames  
+Fixed an issue with the Objective Tracker mover due to a change in Edit Mode  
+Fixed an issue with Macro Text colors in Action Bars due to the ElvUI Skin  
+Fixed Professions missing in bags in Cataclysm due to different Ids  
+Fixed a possible error with the Character Panel in Remix  
+Fixed shadows for detached power  
+Fixed the zhTW locale missing  
 ### 4.1.5
 Added Power bars to the Border Auto Adjust function  
 Added borders to Experience and Reputation databars